+++ app/controllers/api/v1/verses_controller.rb
@@ -2,31 +2,43 @@ class Api::V1::VersesController < ApplicationController
   skip_before_action :verify_authenticity_token
 
   def create
-    verse = Verse.create({
-      lyrics: params[:lyrics],
-      song: Song.find(params[:song_id]),
-      code: detected_language(params[:lyrics])
-    })
+    if current_user
+      verse = Verse.create({
+        lyrics: params[:lyrics],
+        song: Song.find(params[:song_id]),
+        code: detected_language(params[:lyrics])
+      })
 
-    render json: {verse: verse}
+      render json: {verse: verse}
+    else
+      render body: "401 Unauthorized", status: 401
+    end
   end
 
   def update
-    verse = Verse.find(params[:verse_id])
+    if current_user
+      verse = Verse.find(params[:verse_id])
 
-    verse.update({
-      lyrics: params[:lyrics],
-      code: detected_language(params[:lyrics])
-    })
+      verse.update({
+        lyrics: params[:lyrics],
+        code: detected_language(params[:lyrics])
+      })
 
-    render json: {verse: verse}
+      render json: {verse: verse}
+    else
+      render body: "401 Unauthorized", status: 401
+    end
   end
 
   def destroy
-    verse = Verse.find(params[:id])
-    verse.destroy
+    if current_user
+      verse = Verse.find(params[:id])
+      verse.destroy
 
-    render json: {verse: verse}
+      render json: {verse: verse}
+    else
+      render body: "401 Unauthorized", status: 401
+    end
   end
 
   private
